Refactor tier stats into feature-based architecture
Migrates 18 tier-stats files from scattered type-based directories into cohesive feature-based structure under src/features/analysis/tier-stats/ with logical subdirectories (config/, cells/, logic/).
Improves discoverability by grouping related files together and enforces 10-file threshold compliance. Zero logic changes - pure file movement migration verified by passing test suite.
