+++ src/features/analysis/tier-stats/config/tier-stats-config-panel.tsx
@@ -1,15 +1,15 @@
 import { ChevronDown, ChevronUp, RotateCcw, Settings } from 'lucide-react'
-import type { UseTierStatsConfigReturn } from '../hooks/use-tier-stats-config'
-import type { TierStatsAggregation } from '../types/tier-stats-config.types'
-import { getFieldDisplayName, canFieldHaveHourlyRate } from '../utils/tier-stats-config'
-import { useColumnSearch } from '../../settings/column-config/use-column-search'
-import { useColumnReorder } from '../../settings/column-config/use-column-reorder'
-import { SelectedColumnItem } from '../../settings/column-config/selected-column-item'
-import { SearchableColumnPicker } from '../../settings/column-config/searchable-column-picker'
-import { Button } from '../../../components/ui/button'
-import { InfoBox } from '../../../components/ui/info-box'
-import { EmptyState } from '../../../components/ui/empty-state'
-import { FormControl, SelectionButtonGroup } from '../../../components/ui'
+import type { UseTierStatsConfigReturn } from './use-tier-stats-config'
+import type { TierStatsAggregation } from '../types'
+import { getFieldDisplayName, canFieldHaveHourlyRate } from './tier-stats-config-utils'
+import { useColumnSearch } from '@/features/settings/column-config/use-column-search'
+import { useColumnReorder } from '@/features/settings/column-config/use-column-reorder'
+import { SelectedColumnItem } from '@/features/settings/column-config/selected-column-item'
+import { SearchableColumnPicker } from '@/features/settings/column-config/searchable-column-picker'
+import { Button } from '@/components/ui/button'
+import { InfoBox } from '@/components/ui/info-box'
+import { EmptyState } from '@/components/ui/empty-state'
+import { FormControl, SelectionButtonGroup } from '@/components/ui'
 import { getAggregationOptions } from '../logic/tier-stats-aggregation-options'
 
 interface TierStatsConfigPanelProps {
