+++ src/features/analysis/tier-stats/config/tier-stats-persistence.test.ts
@@ -4,9 +4,9 @@ import {
   saveTierStatsConfig,
   clearTierStatsConfig
 } from './tier-stats-persistence'
-import { getDefaultConfig } from './tier-stats-config'
-import type { TierStatsConfig } from '../types/tier-stats-config.types'
-import { TierStatsAggregation } from '../types/tier-stats-config.types'
+import { getDefaultConfig } from './tier-stats-config-utils'
+import type { TierStatsConfig } from '../types'
+import { TierStatsAggregation } from '../types'
 
 describe('tier-stats-persistence', () => {
   const STORAGE_KEY = 'tower-tracking-tier-stats-config'
