+++ src/features/analysis/tier-stats/logic/tier-stats-calculator.ts
@@ -1,15 +1,15 @@
-import type { ParsedGameRun } from '../types/game-run.types'
+import type { ParsedGameRun } from '@/features/data-tracking/types/game-run.types'
 import type {
   DynamicTierStats,
   FieldStats,
   TierStatsColumnConfig,
   TierStatsColumn,
   AvailableField
-} from '../types/tier-stats-config.types'
-import { TierStatsAggregation } from '../types/tier-stats-config.types'
-import { getFieldValue } from './field-utils'
-import { getColumnDisplayName } from './tier-stats-config'
-import { calculateFieldPercentiles } from '../logic/field-percentile-calculation'
+} from '../types'
+import { TierStatsAggregation } from '../types'
+import { getFieldValue } from '@/features/data-tracking/utils/field-utils'
+import { getColumnDisplayName } from '../config/tier-stats-config-utils'
+import { calculateFieldPercentiles } from '@/features/data-tracking/logic/field-percentile-calculation'
 
 /**
  * Calculate dynamic tier stats for all tiers based on selected columns
