+++ src/features/analysis/tier-stats/logic/tier-stats-sort.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect } from 'vitest'
 import { getSortValue, sortTierStats, sortByTier } from './tier-stats-sort'
-import type { DynamicTierStats, TierStatsColumn, FieldStats } from '../types/tier-stats-config.types'
-import type { ParsedGameRun } from '../types/game-run.types'
-import { createGameRunField } from './field-utils'
+import type { DynamicTierStats, TierStatsColumn, FieldStats } from '../types'
+import type { ParsedGameRun } from '@/features/data-tracking/types/game-run.types'
+import { createGameRunField } from '@/features/data-tracking/utils/field-utils'
 
 describe('tier-stats-sort', () => {
   const createMockRun = (tier: number, coins: number, duration: number): ParsedGameRun => ({
