+++ src/features/analysis/tier-stats/tier-stats-table.tsx
@@ -1,14 +1,14 @@
-import { useData } from '../hooks/use-data'
-import { FarmingOnlyIndicator } from './farming-only-indicator'
-import { RunType } from '../types/game-run.types'
-import { TierStatsConfigPanel } from './tier-stats-config-panel'
-import { TierStatsCellTooltip } from './tier-stats-cell-tooltip'
-import { useTierStatsConfig } from '../hooks/use-tier-stats-config'
-import { useDynamicTierStatsTable } from '../hooks/use-dynamic-tier-stats-table'
-import { filterRunsByType } from '../utils/run-type-filter'
-import { getTierStatsCellClassName } from '../utils/tier-stats-cell-styles'
-import { getAggregationDescription } from '../logic/tier-stats-aggregation-options'
-import { LoadingState } from '../../../components/ui/loading-state'
+import { useData } from '@/features/data-tracking/hooks/use-data'
+import { FarmingOnlyIndicator } from '@/features/data-tracking/components/farming-only-indicator'
+import { RunType } from '@/features/data-tracking/types/game-run.types'
+import { TierStatsConfigPanel } from './config/tier-stats-config-panel'
+import { TierStatsCellTooltip } from './cells/tier-stats-cell-tooltip'
+import { useTierStatsConfig } from './config/use-tier-stats-config'
+import { useDynamicTierStatsTable } from './use-dynamic-tier-stats-table'
+import { filterRunsByType } from '@/features/data-tracking/utils/run-type-filter'
+import { getTierStatsCellClassName } from './cells/tier-stats-cell-styles'
+import { getAggregationDescription } from './logic/tier-stats-aggregation-options'
+import { LoadingState } from '@/components/ui/loading-state'
 import * as Tooltip from '@radix-ui/react-tooltip'
 
 export function TierStatsTable() {
