+++ src/features/data-tracking/index.ts
@@ -9,7 +9,7 @@ export { MigrationAlert } from '../settings/data-settings/migration-alert';
 export { RunsTable } from '../game-runs/runs-table';
 export { TabbedRunsTable } from '../game-runs/table/tabbed-runs-table';
 export { DeathsRadarChart } from './components/deaths-radar-chart';
-export { TierStatsTable } from './components/tier-stats-table';
+export { TierStatsTable } from '../analysis/tier-stats/tier-stats-table';
 export { TimeSeriesChart } from './components/time-series-chart';
 export { TierTrendsAnalysis } from '../analysis/tier-trends/tier-trends-analysis';
 export { RunTypeSelector } from './components/run-type-selector';
